+++ packages/mds-ingest-service/service/manager.ts
@@ -37,6 +37,7 @@ export const IngestServiceManager = RpcServer<IngestService, IngestServiceReques
       IngestServiceProvider.getLatestTelemetryForDevices(context, ...args),
     writeEvents: (args, context) => IngestServiceProvider.writeEvents(context, ...args),
     writeEventAnnotations: (args, context) => IngestServiceProvider.writeEventAnnotations(context, ...args),
+    writeTelemetryAnnotations: (args, context) => IngestServiceProvider.writeTelemetryAnnotations(context, ...args),
     getTripEvents: (args, context) => IngestServiceProvider.getTripEvents(context, ...args),
     getEventsWithDeviceAndTelemetryInfoUsingOptions: (args, context) =>
       IngestServiceProvider.getEventsWithDeviceAndTelemetryInfoUsingOptions(context, ...args),