+++ container-images/mds-geography-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @container-images/mds-geography-api
 
+## 0.0.67
+
+### Patch Changes
+
+- @mds-core/mds-geography-api@0.4.19
+
 ## 0.0.66
 
 ### Patch Changes