+++ container-images/mds-transaction-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @container-images/mds-transaction-api
 
+## 0.0.66
+
+### Patch Changes
+
+- @mds-core/mds-transaction-api@0.2.28
+
 ## 0.0.65
 
 ### Patch Changes