+++ packages/mds-agency/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @mds-core/mds-agency
 
+## 0.5.6
+
+### Patch Changes
+
+- 775efc9c: Stream writes telemetry with id
+- Updated dependencies [775efc9c]
+- Updated dependencies [3ca785d0]
+  - @mds-core/mds-ingest-service@0.10.4
+  - @mds-core/mds-db@0.6.11
+  - @mds-core/mds-api-helpers@0.2.19
+
 ## 0.5.5
 
 ### Patch Changes