+++ packages/mds-compliance-api/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @mds-core/mds-compliance-api
 
+## 0.4.30
+
+### Patch Changes
+
+- @mds-core/mds-compliance-service@0.10.25
+- @mds-core/mds-policy-service@0.11.22
+- @mds-core/mds-db@0.6.11
+- @mds-core/mds-test-data@0.4.22
+- @mds-core/mds-api-helpers@0.2.19
+
 ## 0.4.29
 
 ### Patch Changes